restore this example into the source tree
